% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{bio_cache}
\alias{bio_cache}
\alias{setCache}
\alias{removeCache}
\title{Manage cache / download directories for study data}
\usage{
bio_cache(...)

setCache(directory = rappdirs::user_cache_dir("MultiAssayExperimentData"),
  verbose = TRUE, ask = interactive())

removeCache(cancer_study_id)
}
\arguments{
\item{directory}{The file location where the cache is located. Once set
future downloads will go to this folder.}

\item{verbose}{Whether to print descriptive messages}

\item{ask}{logical (default TRUE when interactive session) Confirm the file
location of the cache directory}

\item{cancer_study_id}{A single string from \code{studiesTable} associated
with a study tarball}
}
\description{
Managing data downloads is important to save disk space and
re-downloading data files. This can be done effortlessly via the integrated
BiocFileCache system.
}
\section{bio_cache}{

Get the directory location of the cache. It will prompt the user to create
a cache if not already created. A specific directory can be used via
\code{setCache}.
}

\section{setCache}{

Specify the directory location of the data cache. By default, it will
got to the user's home/.cache and "appname" directory as specified by
\link{user_cache_dir}. (default appname: MultiAssayExperimentData)
}

\section{removeCache}{

Some files may become corrupt when downloading, this function allows
the user to delete the tarball associated with a \code{cancer_study_id} in the
cache.
}

