% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cBio_cache}
\alias{cBio_cache}
\alias{removeCache}
\alias{setCache}
\alias{cBio_cache}
\title{Manage cache / download directories for study data}
\usage{
removeCache(cancer_study_id)

setCache(directory = rappdirs::user_cache_dir("MultiAssayExperimentData"),
  verbose = TRUE, ask = interactive())

cBio_cache()
}
\arguments{
\item{cancer_study_id}{A single string from \code{studiesTable} associated
with a study tarball}

\item{directory}{The file location where the cache is located. Once set
future downloads will go to this folder.}
}
\description{
Managing data downloads is important to save disk space and
re-downloading data files. This can be done effortlessly via the integrated
BiocFileCache system.
}
\section{setCache}{

Specify the directory location of the data cache. By default, it will
got to the user's home directory and project directory as specified by
\link{user_cache_dir}
}

\section{removeCache}{

Some files may become corrupt when downloading, this function allows
the user to delete the tarball associated with a `cancer_study_id``
}

