% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{clinicalData}
\alias{clinicalData}
\title{Obtain clinical data}
\usage{
clinicalData(cbio, studyId = "acc_tcga")
}
\arguments{
\item{cbio}{An object of class `cBioPortal`}

\item{studyId}{A single string indicating the "studyId" as taken from
`getStudies`}
}
\description{
Obtain clinical data
}
