% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadcBioPortal.R
\name{downloadcBioPortal}
\alias{downloadcBioPortal}
\title{Download and cache study dataset}
\usage{
downloadcBioPortal(cancer_study_id, use_cache = TRUE)
}
\arguments{
\item{cancer_study_id}{The cBioPortal study identifier}

\item{use_cache}{logical (default TRUE) create the default cache location
and use it to track downloaded data. If data found in the cache, data will
not be re-downloaded. A path can also be provided to data cache location.}
}
\description{
Provide a `cancer_study_id` from the `studiesTable` and retrieve
the study tarball from MSKCC's cBioPortal
}
\examples{

downloadcBioPortal("laml_tcga_pub", use_cache = tempdir())

}
