% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{geneTable}
\alias{geneTable}
\title{Get a table of all genes}
\usage{
geneTable(cbio, ...)
}
\arguments{
\item{cbio}{An object of class `cBioPortal`}
}
\description{
Query the API to get a list of Entrez Gene IDs, Hugo symbols, etc.
}
