% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{getStudies}
\alias{getStudies}
\title{Obtain a table of studies and associated metadata}
\usage{
getStudies(cbio)
}
\arguments{
\item{cbio}{An object of class `cBioPortal`}
}
\description{
Obtain a table of studies and associated metadata
}
