% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importcBioPortal.R
\name{importcBioPortal}
\alias{importcBioPortal}
\title{Title importcBioPortal: convert a .tar.gz file downloaded from
http://www.cbioportal.org/data_sets.jsp to a MultiAssayExperiment object}
\usage{
importcBioPortal(cancer_study_id, cancer_file = NULL,
  dir_location = tempdir(), split.field = c("Tumor_Sample_Barcode", "ID"),
  names.field = c("Hugo_Symbol", "Entrez_Gene_Id"))
}
\arguments{
\item{cancer_study_id}{The cBioPortal study identifier}

\item{cancer_file}{The (optional) location of a previously downloaded tar
file}

\item{dir_location}{If no tar.gz file provided, the location to download
and untar the study files}

\item{split.field}{A character vector of possible column names for the column
that is used to identify samples in a mutations or copy number file.}

\item{names.field}{A character vector of possible column names for the column
that is used to label ranges from a mutations or copy number file.}
}
\value{
A \code{MultiAssayExperiment} object
}
\description{
Title importcBioPortal: convert a .tar.gz file downloaded from
http://www.cbioportal.org/data_sets.jsp to a MultiAssayExperiment object
}
\examples{

data(studiesTable)

(laml <- studiesTable[["cancer_study_id"]][3L])
mae <- importcBioPortal(laml)

}
\author{
Levi Waldron, M. Ramos
}
