% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{molecularProfiles}
\alias{molecularProfiles}
\title{Produce molecular profiles dataset}
\usage{
molecularProfiles(cbio, studyId = "acc_tcga", projection = c("SUMMARY",
  "ID", "DETAILED", "META"))
}
\arguments{
\item{cbio}{An object of class `cBioPortal`}

\item{studyId}{A single string indicating the "studyId" as taken from
`getStudies`}
}
\description{
Produce molecular profiles dataset
}
