% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{molecularSlice}
\alias{molecularSlice}
\title{Produce small dataset of molecular profile data}
\usage{
molecularSlice(cbio, profileId = "acc_tcga_rna_seq_v2_mrna",
  entrezGeneIds = c(1, 2), sampleIds = c("TCGA-OR-A5J1-01",
  "TCGA-OR-A5J2-01"))
}
\arguments{
\item{cbio}{An object of class `cBioPortal`}

\item{profileId}{A single string indicating molecular profile ID}

\item{entrezGeneIds}{A numeric vector indicating entrez gene IDs}

\item{sampleIds}{A character vector for TCGA sample identifiers}
}
\description{
This function will query the `fetchAllMolecularDataInMolecularProfileUsingPOST`
endpoint to obtain data
}
\examples{

cc <- cBioPortal()
molecularSlice(cc)

}
