% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCache.R
\name{removeCache}
\alias{removeCache}
\title{Delete downloaded study tarballs from cache}
\usage{
removeCache(cancer_study_id)
}
\arguments{
\item{cancer_study_id}{A single string from \code{studiesTable} associated
with a study tarball}
}
\description{
Some files may become corrupt when downloading, this function allows
the user to delete the tarball associated with a `cancer_study_id``
}
