% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{sampleLists}
\alias{sampleLists}
\title{Provide a sample lists within a study}
\usage{
sampleLists(cbio, studyId = "acc_tcga")
}
\arguments{
\item{cbio}{An object of class `cBioPortal`}

\item{studyId}{A single string indicating the "studyId" as taken from
`getStudies`}
}
\description{
Given a particular `studyId``, this function will return all the available
`sampleListId` identifiers
}
