% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{samplesInSampleLists}
\alias{samplesInSampleLists}
\title{Return all samples within sample lists}
\usage{
samplesInSampleLists(cbio, sampleListIds = c("acc_tcga_all"))
}
\arguments{
\item{cbio}{An object of class `cBioPortal`}

\item{sampleListIds}{A character vector of sampleListId as obtained from
sampleLists}
}
\description{
Provide a sampleListId and the function will return a CharacterList of
associated TCGA barcodes
}
\examples{
samplesInSampleLists(cc, "acc_tcga_rppa")

}
