% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cBioCache}
\alias{cBioCache}
\alias{setCache}
\alias{removePackCache}
\alias{removeDataCache}
\title{Manage cache / download directories for study data}
\usage{
cBioCache(...)

setCache(
  directory = tools::R_user_dir("cBioPortalData", "cache"),
  verbose = TRUE,
  ask = interactive()
)

removePackCache(cancer_study_id, dry.run = TRUE)

removeDataCache(
  api,
  studyId = NA_character_,
  genePanelId = NA_character_,
  genes = NA_character_,
  molecularProfileIds = NULL,
  sampleListId = NULL,
  by = c("entrezGeneId", "hugoGeneSymbol"),
  dry.run = TRUE,
  ...
)
}
\arguments{
\item{...}{For \code{cBioCache}, arguments passed to \code{setCache}}

\item{directory}{The file location where the cache is located. Once set
future downloads will go to this folder.}

\item{verbose}{Whether to print descriptive messages}

\item{ask}{logical (default TRUE when interactive session) Confirm the file
location of the cache directory}

\item{cancer_study_id}{A single string from \code{studiesTable} associated
with a study tarball}

\item{dry.run}{logical Whether or not to remove cache files (default TRUE).}

\item{api}{An API object of class `cBioPortal` from the `cBioPortal`
function}

\item{studyId}{character(1) Indicates the "studyId" as taken from
`getStudies`}

\item{genePanelId}{character(1) Identifies the gene panel, as obtained
from the `genePanels` function}

\item{genes}{character() Either Entrez gene identifiers or Hugo gene
symbols. When included, the 'by' argument indicates the type of
identifier provided and 'genePanelId' is ignored. Preference is
given to Entrez IDs due to faster query responses.}

\item{molecularProfileIds}{character() A vector of molecular profile IDs}

\item{sampleListId}{character(1) A sample list identifier as obtained from
`sampleLists()``}

\item{by}{character(1) Either 'entrezGeneId' or 'hugoGeneSymbol' for row
metadata (default: 'entrezGeneId')}
}
\value{
cBioCache: The path to the cache location
}
\description{
Managing data downloads is important to save disk space and
re-downloading data files. This can be done effortlessly via the integrated
BiocFileCache system.
}
\section{cBioCache}{

Get the directory location of the cache. It will prompt the user to create
a cache if not already created. A specific directory can be used via
\code{setCache}.
}

\section{setCache}{

Specify the directory location of the data cache. By default, it will
go to the user directory as given by:
\preformatted{
    tools::R_user_dir("cBioPortalData", "cache")
}
}

\section{removePackCache}{

Some files may become corrupt when downloading, this function allows
the user to delete the tarball associated with a \code{cancer_study_id} in the
cache. This only works for the \code{cBioDataPack} function. To remove the entire
\code{cBioPortalData} cache, run \code{unlink("~/.cache/cBioPortalData")}.
}

\examples{

cBioCache()

removePackCache("acc_tcga", dry.run = TRUE)


cbio <- cBioPortal()

cBioPortalData(
    cbio, by = "hugoGeneSymbol",
    studyId = "acc_tcga",
    genePanelId = "AmpliSeq",
    molecularProfileIds =
        c("acc_tcga_rppa", "acc_tcga_linear_CNA", "acc_tcga_mutations")
)

removeDataCache(
    cbio, by = "hugoGeneSymbol",
    studyId = "acc_tcga",
    genePanelId = "AmpliSeq",
    molecularProfileIds =
        c("acc_tcga_rppa", "acc_tcga_linear_CNA", "acc_tcga_mutations"),
    dry.run = TRUE
)

}
