% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioDataPack.R
\name{cBioDataPack}
\alias{cBioDataPack}
\title{Obtain pre-packaged data from cBioPortal and represent as
a MultiAssayExperiment object}
\usage{
cBioDataPack(cancer_study_id, use_cache = TRUE,
  split.field = c("Tumor_Sample_Barcode", "ID"),
  names.field = c("Hugo_Symbol", "Entrez_Gene_Id", "Gene"))
}
\arguments{
\item{cancer_study_id}{The cBioPortal study identifier as in
\url{https://cbioportal.org/webAPI}}

\item{use_cache}{logical (default TRUE) create the default cache location
and use it to track downloaded data. If data found in the cache, data will
not be re-downloaded. A path can also be provided to data cache location.}

\item{split.field}{A character vector of possible column names for the column
that is used to identify samples in a mutations or copy number file.}

\item{names.field}{A character vector of possible column names for the column
that is used to label ranges from a mutations or copy number file.}
}
\value{
A \linkS4class{MultiAssayExperiment} object
}
\description{
The \code{cBioDataPack} function allows the user to
download and process cancer study datasets found in MSKCC's cBioPortal.
Output datasets use the \linkS4class{MultiAssayExperiment} data
representation to faciliate analysis and data management operations.
}
\details{
The list of datasets can be found in the `studiesTable` dataset
by doing \code{data("studiesTable")}. Some datasets may not be available
for download. Please refer to the
\href{http://cbioportal.org/data_sets.jsp}{website} for the full list of
datasets
}
\examples{

data(studiesTable)

(laml <- studiesTable[["cancer_study_id"]][3L])

mae <- cBioDataPack(laml)

}
\seealso{
\url{http://cbioportal.org/data_sets.jsp}
}
\author{
Levi Waldron, M. Ramos
}
