% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\docType{class}
\name{cBioPortal-class}
\alias{cBioPortal-class}
\alias{.cBioPortal}
\title{A class for representing the cBioPortal API}
\description{
The cBioPortal class is a product of the cBioPortal API that
    inherits from the `Service` class
}
\examples{

cbio <- cBioPortal()

}
\seealso{
AnVIL::Service
}
