% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\docType{class}
\name{cBioPortal-class}
\alias{cBioPortal-class}
\alias{.cBioPortal}
\title{A class for representing the cBioPortal API protocol}
\description{
The `cBioPortal` class is a representation of the cBioPortal
    API protocol that directly inherits from the `Service` class in the
    `AnVIL` package
}
\examples{

cBioPortal()

}
\seealso{
\link{cBioPortal}, \link[AnVIL]{Service}
}
