% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{cBioPortal}
\alias{cBioPortal}
\title{API Entry function for the cBioPortal data service}
\usage{
cBioPortal()
}
\value{
An object of class 'cBioPortal'
}
\description{
This function allows the use of the cBioPortal API
}
