% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortalData.R
\name{cBioPortalData}
\alias{cBioPortalData}
\title{Download data from the cBioPortal API}
\usage{
cBioPortalData(
  api,
  studyId = NA_character_,
  genePanelId = NA_character_,
  molecularProfileIds = NULL,
  sampleListId = NULL,
  by = c("entrezGeneId", "hugoGeneSymbol")
)
}
\arguments{
\item{api}{An API object of class `cBioPortal` from the `cBioPortal`
function}

\item{studyId}{character(1) Indicates the "studyId" as taken from
`getStudies`}

\item{genePanelId}{character(1) Identifies the gene panel, as obtained
from the `genePanels` function}

\item{molecularProfileIds}{character() A vector of molecular profile IDs}

\item{sampleListId}{character(1) A sample list identifier as obtained from
`sampleLists()``}

\item{by}{character(1) Either 'entrezGeneId' or 'hugoGeneSymbol' for row
metadata}
}
\value{
A \linkS4class{MultiAssayExperiment} object
}
\description{
Obtain a \code{MultiAssayExperiment} object for a particular gene panel,
\code{studyId}, \code{molecularProfileIds}, and \code{sampleListIds} combination. Default
\code{molecularProfileIds} and \code{sampleListIds} are set to NULL for including all
data. This option is best for users who wish to obtain a section of the
study data that pertains to a specific molecular profile and gene panel
combination. For users looking to download the entire study data as provided
by the \url{https://cbioportal.org/datasets}, refer to \code{cBioDataPack}.
}
\examples{

cbio <- cBioPortal()

cBioPortalData(cbio, by = "hugoGeneSymbol", studyId = "acc_tcga",
    genePanelId = "IMPACT341",
    molecularProfileIds = c("acc_tcga_rppa", "acc_tcga_linear_CNA")
    )

}
\seealso{
\link{cBioDataPack}
}
