% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{cBioPortalData}
\alias{cBioPortalData}
\title{Download data from the cBioPortal API}
\usage{
cBioPortalData(cbio, by = c("entrezGeneId", "hugoGeneSymbol"),
  genePanelId = "bait_v5", studyId = "acc_tcga",
  molecularProfileIds = c("acc_tcga_rppa", "acc_tcga_rppa_Zscores",
  "acc_tcga_gistic"), sampleListId = NULL)
}
\arguments{
\item{cbio}{An object of class `cBioPortal`}

\item{by}{character(1) Whether to use 'entrezGeneId' or 'hugoGeneSymbol'
as row metadata}

\item{genePanelId}{A character vector identifying a gene panel, as obtained
from the `genePanels` function}

\item{studyId}{A single string indicating the "studyId" as taken from
`getStudies`}

\item{molecularProfileIds}{A character vector of molecular profile IDs}
}
\description{
Obtain a `MultiAssayExperiment` object for a particular gene panel,
`studyId`, `molecularProfileIds`, and `sampleListIds` combination. Default
`molecularProfileIds` and `sampleListIds` are set to NULL for including all
data.
}
\examples{

cb <- cBioPortal()
cBioPortalData(cb, by = "hugoGeneSymbol")

}
