% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortalData.R
\name{cBioPortalData}
\alias{cBioPortalData}
\title{Download data from the cBioPortal API}
\usage{
cBioPortalData(
  api,
  studyId = NA_character_,
  genePanelId = NA_character_,
  genes = NA_character_,
  molecularProfileIds = NULL,
  sampleListId = NULL,
  by = c("entrezGeneId", "hugoGeneSymbol")
)
}
\arguments{
\item{api}{An API object of class `cBioPortal` from the `cBioPortal`
function}

\item{studyId}{character(1) Indicates the "studyId" as taken from
`getStudies`}

\item{genePanelId}{character(1) Identifies the gene panel, as obtained
from the `genePanels` function}

\item{genes}{character() Either Entrez gene identifiers or Hugo gene
symbols. When included, the 'by' argument indicates the type of
identifier provided and 'genePanelId' is ignored. Preference is
given to Entrez IDs due to faster query responses.}

\item{molecularProfileIds}{character() A vector of molecular profile IDs}

\item{sampleListId}{character(1) A sample list identifier as obtained from
`sampleLists()``}

\item{by}{character(1) Either 'entrezGeneId' or 'hugoGeneSymbol' for row
metadata (default: 'entrezGeneId')}
}
\value{
A \linkS4class{MultiAssayExperiment} object
}
\description{
Obtain a \code{MultiAssayExperiment} object for a particular gene panel,
\code{studyId}, \code{molecularProfileIds}, and \code{sampleListIds} combination. Default
\code{molecularProfileIds} and \code{sampleListIds} are set to NULL for including all
data. This option is best for users who wish to obtain a section of the
study data that pertains to a specific molecular profile and gene panel
combination. For users looking to download the entire study data as provided
by the \url{https://cbioportal.org/datasets}, refer to \code{cBioDataPack}.
}
\details{
As of May 2020, there were about 96.6 percent of the 268 datasets
successfully imported. The datasets that currently fail to import are:
\preformatted{
        c("all_stjude_2015", "sclc_ucologne_2015", "skcm_ucla_201
        "sclc_jhu", "gbm_tcga_pub2013", "hnsc_tcga_pub", "kirc_tc
        "brca_tcga_pub", "brca_tcga_pub2015")
    }
Note that changes to the cBioPortal API may affect this rate at any
time. If you encounter any issues, please open a GitHub issue at the
\url{https://github.com/waldronlab/cBioPortalData/issues/} page with
a fully reproducible example.
}
\examples{

cbio <- cBioPortal()

samps <- samplesInSampleLists(cbio, "acc_tcga_rppa")[[1]]

getGenePanelMolecular(
    cbio, molecularProfileIds = c("acc_tcga_rppa", "acc_tcga_linear_CNA"),
    samps
)

acc_tcga <- cBioPortalData(
    cbio, by = "hugoGeneSymbol",
    studyId = "acc_tcga",
    genePanelId = "AmpliSeq",
    molecularProfileIds =
        c("acc_tcga_rppa", "acc_tcga_linear_CNA", "acc_tcga_mutations")
)

}
\seealso{
\link{cBioDataPack}
}
