% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{cBioPortalData}
\alias{cBioPortalData}
\title{Download data from the cBioPortal API}
\usage{
cBioPortalData(api, studyId = NA_character_,
  genePanelId = NA_character_, molecularProfileIds = NULL,
  sampleListId = NULL, by = c("entrezGeneId", "hugoGeneSymbol"),
  idConvert = identity)
}
\arguments{
\item{api}{An API object of class `cBioPortal` from the `cBioPortal`
function}

\item{studyId}{character(1) Indicates the "studyId" as taken from
`getStudies`}

\item{genePanelId}{character(1) Identifies the gene panel, as obtained
from the `genePanels` function}

\item{molecularProfileIds}{character() A vector of molecular profile IDs}

\item{sampleListId}{character(1) A sample list identifier as obtained from
`sampleLists()``}

\item{by}{character(1) Either 'entrezGeneId' or 'hugoGeneSymbol' for row
metadata}

\item{idConvert}{function(default: `identity()`) A function to process
identifiers for matching patients to samples. It catches TCGA samples
automatically and uses `TCGAutils::TCGAbarcode` instead.}
}
\description{
Obtain a `MultiAssayExperiment` object for a particular gene panel,
`studyId`, `molecularProfileIds`, and `sampleListIds` combination. Default
`molecularProfileIds` and `sampleListIds` are set to NULL for including all
data.
}
\examples{

cbio <- cBioPortal()

cBioPortalData(cbio, by = "hugoGeneSymbol", studyId = "acc_tcga",
    genePanelId = "IMPACT341",
    molecularProfileIds = c("acc_tcga_rppa", "acc_tcga_linear_CNA")
)

}
