% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBioPortal.R
\name{cBioPortalData}
\alias{cBioPortalData}
\title{Download data from the cBioPortal API}
\usage{
cBioPortalData(cbio, studyId, genePanelId, molecularProfileIds = NULL,
  sampleListId = NULL, by = c("entrezGeneId", "hugoGeneSymbol"))
}
\arguments{
\item{cbio}{An object of class `cBioPortal`}

\item{studyId}{character(1) Indicates the "studyId" as taken from
`getStudies`}

\item{genePanelId}{character(1) Identifies the gene panel, as obtained
from the `genePanels` function}

\item{molecularProfileIds}{character() A vector of molecular profile IDs}

\item{by}{character(1) Whether to use 'entrezGeneId' or 'hugoGeneSymbol'
as row metadata}
}
\description{
Obtain a `MultiAssayExperiment` object for a particular gene panel,
`studyId`, `molecularProfileIds`, and `sampleListIds` combination. Default
`molecularProfileIds` and `sampleListIds` are set to NULL for including all
data.
}
\examples{

cb <- cBioPortal()
cBioPortalData(cb, by = "hugoGeneSymbol")

}
