% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importcBioPortal.R
\name{importcBioPortal}
\alias{importcBioPortal}
\title{Convert a data file downloaded from MSKCC's cBioPortal to
a MultiAssayExperiment object}
\usage{
importcBioPortal(cancer_study_id, use_cache = TRUE,
  split.field = c("Tumor_Sample_Barcode", "ID"),
  names.field = c("Hugo_Symbol", "Entrez_Gene_Id", "Gene"))
}
\arguments{
\item{cancer_study_id}{The cBioPortal study identifier}

\item{use_cache}{logical (default TRUE) if data found in the cache, do
not download again, otherwise re-download data}

\item{split.field}{A character vector of possible column names for the column
that is used to identify samples in a mutations or copy number file.}

\item{names.field}{A character vector of possible column names for the column
that is used to label ranges from a mutations or copy number file.}
}
\value{
A \code{MultiAssayExperiment} object
}
\examples{

data(studiesTable)

(laml <- studiesTable[["cancer_study_id"]][3L])

mae <- importcBioPortal(laml)

}
\seealso{
\url{http://cbioportal.org/data_sets.jsp}
}
\author{
Levi Waldron, M. Ramos
}
